Rename project to Smart Moving Reloaded
After some deliberation, I realized that with mixins there isn't much of a point to having a separate API for interacting with the player classes. Other mods can use mixins to interact with the player classes if need be (and I don't have to maintain a whole API). Also substantially reduces the size of the codebase.